refactor: flattened directory structure
