+++ Makefile
@@ -2,4 +2,8 @@ lint-fix:
 	gofmt -d -w .
 
 lint: 
-	gofmt -d
+	gofmt -d .
+
+test:
+	go test ./core
+	go test ./core -bench=.
